Bug 2042178 - Convert maxDataBytes to Int in IPProtectionScreen UI r=mavduevskiy,android-reviewers,android-l10n-reviewers,flod
Differential Revision: https://phabricator.services.mozilla.com/D302578

UltraBlame original commit: 832c1d8eb8c51080b5927097da176d98cd5a1320

+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/settings/IPProtectionScreen.kt
@@ -154,7 +154,7 @@ private fun DataLimitSection(
             )
         } else {
             Text(
-                text = stringResource(R.string.ip_protection_data_limit_reached_description, state.maxDataGb),
+                text = stringResource(R.string.ip_protection_data_limit_reached_description, state.maxDataGb.toInt()),
                 style = FirefoxTheme.typography.body2,
                 color = MaterialTheme.colorScheme.error,
             )
